+++ crates/macros/src/config/config.rs
@@ -1,5 +1,4 @@
-use crate::common::{Container, TaggedFormat};
-use crate::common_schema::*;
+use crate::common::{Container, ContainerSerdeArgs, SerdeMeta, TaggedFormat};
 use darling::FromDeriveInput;
 use proc_macro2::{Ident, TokenStream};
 use quote::{format_ident, quote, ToTokens};
