+++ crates/macros/src/lib.rs
@@ -1,5 +1,4 @@
 mod common;
-mod common_schema;
 mod config;
 mod config_enum;
 mod utils;
